% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saif.R
\name{bail_out}
\alias{bail_out}
\title{Bail out on an event}
\usage{
bail_out(category = NULL)
}
\arguments{
\item{category}{Optional. Category of excuse to generate.}
}
\value{
A string containing the excuse.
}
\description{
Generate a random excuse that Saif could have said himself.
If a category is given, generate an excuse for that category specifically.
}
\examples{
exc <- bail_out("sport")
sprintf("I can't because \%s", exc)
}
