% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excuses.R
\docType{data}
\name{excuses}
\alias{excuses}
\title{Excuses}
\format{
An object of the class \code{"data.frame"}
\describe{
  \item{sentence}{A string containing the excuse.}
  \item{category}{A string defining the category.}
}
}
\usage{
data(excuses)
}
\description{
Excuses
}
\details{
All the excuses are of the form "I + {verb} + {complement}"
}
\keyword{datasets}
