% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excuses.R
\docType{data}
\name{excuses}
\alias{excuses}
\title{Excuses}
\format{
An object of the class \code{"data.frame"}
\describe{
  \item{sentence}{A string containing the excuse.}
  \item{category}{A string defining the category.}
}
}
\usage{
data(excuses)
}
\description{
List of excuses that Saif has already used or that he could use in any kind
of situation, along with their category of use.
}
\keyword{datasets}
